refactor(jellyfin): move the command ingress into the product service
handle_command owns Play/Stop/Pause/Resume/Seek/Next/volume/mute ingress
in jellyfin_service, with route-side seams injected: playback control
dispatch and UI eventing come in as late-binding adapters built by the
routes wrapper, so monkeypatches keep intercepting. Dedupe state, the
suppression helpers, and smart_item_from_url moved with it. The play
path's now-playing write goes through playback_service.update_now_playing
and routes/__init__.py leaves the QUEUE writer set entirely.
