+++ app/relaytv_app/integrations/jellyfin_service.py
@@ -16,15 +16,17 @@ from __future__ import annotations
 import os
 import re
 import threading
+import time
 
 from fastapi import HTTPException
 
 from typing import Protocol
 from urllib.parse import parse_qsl, urlencode, urlsplit, urlunsplit
 
-from .. import playback_service, player, state, video_profile
+from .. import playback_service, player, resolver, state, video_profile
 from ..config import runtime_config
 from ..debug import get_logger
+from ..thumb_cache import attach_local_thumbnail
 from . import jellyfin_receiver
 
 logger = get_logger("jellyfin")
@@ -2227,6 +2229,789 @@ def progress_snapshot() -> dict | None:
         except Exception:
             pass
     return payload
+
+
+# Command ingress dedupe state (play debounce, command-id dedupe, UI action
+# dedupe). Reset on connect/disconnect via reset_command_state().
+_JELLYFIN_PLAY_DEBOUNCE_LOCK = threading.Lock()
+_JELLYFIN_LAST_PLAY: dict[str, object] = {"ts": 0.0, "url": "", "item_id": "", "start_pos": None}
+_JELLYFIN_COMMAND_DEDUPE_LOCK = threading.Lock()
+_JELLYFIN_RECENT_COMMAND_IDS: dict[str, float] = {}
+_JELLYFIN_UI_ACTION_DEDUPE_LOCK = threading.Lock()
+_JELLYFIN_LAST_UI_ACTION: dict[str, object] = {"ts": 0.0, "command": "", "item_id": "", "resume_pos": None}
+
+
+def _extract_api_key_from_url(url: str) -> str:
+    try:
+        q = dict(parse_qsl(urlsplit(str(url or "").strip()).query, keep_blank_values=True))
+        return str(q.get("api_key") or q.get("ApiKey") or "").strip()
+    except Exception:
+        return ""
+
+
+def smart_item_from_url(url: str, *, start_pos: float | None = None, lightweight: bool = False) -> dict:
+    """
+    Build a playback item for smart/jellyfin paths.
+    If the URL looks like Jellyfin media, enrich title/thumbnail/resume from Jellyfin APIs.
+    """
+    shared = str(url or "").strip()
+    item_id = jellyfin_receiver.extract_item_id_from_url(shared)
+    st = jellyfin_receiver.status()
+    if item_id and (st.get("server_url") or looks_like_media_url(shared)):
+        origin = url_origin(shared)
+        server_url = origin or str(st.get("server_url") or "")
+        link_api_key = _extract_api_key_from_url(shared)
+        token = link_api_key or access_token()
+        pref_audio_idx, pref_sub_idx = preferred_stream_indices(item_id)
+        normalized_url = normalize_source_url(shared, server_url=server_url, api_key=token)
+        normalized_url = apply_stream_params(
+            normalized_url,
+            audio_stream_index=pref_audio_idx,
+            subtitle_stream_index=pref_sub_idx,
+        )
+        # Share links commonly use /Items/<id>/Download - convert to stream endpoint for consistency.
+        low_path = (urlsplit(shared).path or "").lower()
+        if "/items/" in low_path and "/download" in low_path:
+            normalized_url = build_item_stream_url(
+                item_id,
+                server_url=server_url,
+                api_key=token,
+                audio_stream_index=pref_audio_idx,
+                subtitle_stream_index=pref_sub_idx,
+            )
+        media_source_id = extract_media_source_id_from_url(normalized_url)
+        try:
+            settings_snapshot = state.get_settings()
+        except Exception:
+            settings_snapshot = {}
+        selected = select_playback_url(
+            item_id=item_id,
+            source_url=normalized_url,
+            server_url=server_url,
+            api_key=token,
+            media_source_id=media_source_id,
+            audio_stream_index=pref_audio_idx,
+            subtitle_stream_index=pref_sub_idx,
+            settings=settings_snapshot,
+        )
+        normalized_url = normalize_source_url(
+            str(selected.get("url") or normalized_url),
+            server_url=server_url,
+            api_key=token,
+        )
+        media_source_id = _first_nonempty_str(
+            [
+                str(selected.get("media_source_id") or "").strip(),
+                extract_media_source_id_from_url(normalized_url),
+                media_source_id,
+            ]
+        )
+        item: dict[str, object] = {
+            "url": normalized_url,
+            "provider": "jellyfin",
+            "jellyfin_item_id": item_id,
+            **({"jellyfin_media_source_id": media_source_id} if media_source_id else {}),
+            "jellyfin_stream_mode": str(selected.get("mode") or "direct"),
+            "jellyfin_stream_reason": str(selected.get("reason") or ""),
+        }
+        meta = jellyfin_receiver.get_item_metadata(item_id, token_override=token, server_url_override=server_url)
+        if isinstance(meta, dict):
+            title = str(meta.get("title") or "").strip()
+            channel = str(meta.get("channel") or "").strip()
+            thumb = str(meta.get("thumbnail") or "").strip()
+            if title:
+                item["title"] = title
+            if channel:
+                item["channel"] = channel
+            if thumb:
+                item["thumbnail"] = thumb
+            if start_pos is None:
+                try:
+                    rp = meta.get("resume_pos")
+                    if rp is not None:
+                        item["resume_pos"] = float(rp)
+                except Exception:
+                    pass
+        if start_pos is not None:
+            try:
+                item["resume_pos"] = float(start_pos)
+            except Exception:
+                pass
+        return attach_local_thumbnail(item)
+    # Non-Jellyfin path: existing resolver behavior.
+    try:
+        return resolver.make_item(shared, lightweight=lightweight)
+    except TypeError:
+        # Compatibility for tests/patches that mock make_item(url) without kwargs.
+        return resolver.make_item(shared)
+
+
+def reset_command_state() -> None:
+    with _JELLYFIN_PLAY_DEBOUNCE_LOCK:
+        _JELLYFIN_LAST_PLAY.update({"ts": 0.0, "url": "", "item_id": "", "start_pos": None})
+    with _JELLYFIN_COMMAND_DEDUPE_LOCK:
+        _JELLYFIN_RECENT_COMMAND_IDS.clear()
+    with _JELLYFIN_UI_ACTION_DEDUPE_LOCK:
+        _JELLYFIN_LAST_UI_ACTION.update({"ts": 0.0, "command": "", "item_id": "", "resume_pos": None})
+
+
+def is_duplicate_command(command_id: str) -> bool:
+    cid = str(command_id or "").strip()
+    if not cid:
+        return False
+    ttl = max(1.0, float(os.getenv("RELAYTV_JELLYFIN_COMMAND_ID_TTL_SEC", "30")))
+    now_ts = time.time()
+    with _JELLYFIN_COMMAND_DEDUPE_LOCK:
+        # prune expired ids
+        expired = [k for k, ts in _JELLYFIN_RECENT_COMMAND_IDS.items() if (now_ts - ts) > ttl]
+        for k in expired:
+            _JELLYFIN_RECENT_COMMAND_IDS.pop(k, None)
+        if cid in _JELLYFIN_RECENT_COMMAND_IDS:
+            _JELLYFIN_RECENT_COMMAND_IDS[cid] = now_ts
+            return True
+        _JELLYFIN_RECENT_COMMAND_IDS[cid] = now_ts
+        return False
+
+
+def should_suppress_duplicate_play(url: str, item_id: str, start_pos: float | None) -> bool:
+    window_sec = max(0.0, float(os.getenv("RELAYTV_JELLYFIN_PLAY_DEBOUNCE_SEC", "1.5")))
+    if window_sec <= 0:
+        return False
+    now_ts = time.time()
+    with _JELLYFIN_PLAY_DEBOUNCE_LOCK:
+        last_ts = float(_JELLYFIN_LAST_PLAY.get("ts") or 0.0)
+        if now_ts - last_ts > window_sec:
+            _JELLYFIN_LAST_PLAY.update({"ts": now_ts, "url": url, "item_id": item_id, "start_pos": start_pos})
+            return False
+        same_url = str(_JELLYFIN_LAST_PLAY.get("url") or "") == str(url or "")
+        same_item = str(_JELLYFIN_LAST_PLAY.get("item_id") or "") == str(item_id or "")
+        last_start = _JELLYFIN_LAST_PLAY.get("start_pos")
+        try:
+            delta = abs(float(last_start) - float(start_pos)) if (last_start is not None and start_pos is not None) else 0.0
+        except Exception:
+            delta = 0.0
+        same_start = (last_start is None and start_pos is None) or (delta < 1.0)
+        suppressed = same_url and (same_item or (not item_id)) and same_start
+        _JELLYFIN_LAST_PLAY.update({"ts": now_ts, "url": url, "item_id": item_id, "start_pos": start_pos})
+        return suppressed
+
+
+def should_suppress_duplicate_ui_action(command: str, item_id: str, resume_pos: float | None) -> bool:
+    window_sec = max(0.0, float(os.getenv("RELAYTV_JELLYFIN_UI_ACTION_DEDUPE_SEC", "1.5")))
+    if window_sec <= 0:
+        return False
+    now_ts = time.time()
+    norm_cmd = str(command or "").strip().lower()
+    norm_item_id = canonical_item_id(item_id)
+    with _JELLYFIN_UI_ACTION_DEDUPE_LOCK:
+        last_ts = float(_JELLYFIN_LAST_UI_ACTION.get("ts") or 0.0)
+        if now_ts - last_ts > window_sec:
+            _JELLYFIN_LAST_UI_ACTION.update(
+                {"ts": now_ts, "command": norm_cmd, "item_id": norm_item_id, "resume_pos": resume_pos}
+            )
+            return False
+        same_cmd = str(_JELLYFIN_LAST_UI_ACTION.get("command") or "") == norm_cmd
+        same_item = str(_JELLYFIN_LAST_UI_ACTION.get("item_id") or "") == norm_item_id
+        last_resume = _JELLYFIN_LAST_UI_ACTION.get("resume_pos")
+        try:
+            delta = abs(float(last_resume) - float(resume_pos)) if (last_resume is not None and resume_pos is not None) else 0.0
+        except Exception:
+            delta = 0.0
+        same_resume = (last_resume is None and resume_pos is None) or (delta < 1.0)
+        suppressed = same_cmd and same_item and same_resume
+        _JELLYFIN_LAST_UI_ACTION.update(
+            {"ts": now_ts, "command": norm_cmd, "item_id": norm_item_id, "resume_pos": resume_pos}
+        )
+        return suppressed
+
+
+def handle_command(req: CommandReqLike, *, controls: dict, ui: dict):
+    """Normalized Jellyfin command ingress (v1: Play/Stop/Pause/Resume/Seek/Next).
+
+    ``controls`` maps stop/pause/resume/seek/next/previous/set_volume/mute
+    to route-side playback control callables; ``ui`` provides toast,
+    notification_display_sec, queue_event, and jellyfin_event callbacks.
+    Route-facing side effects stay behind these seams so the service never
+    imports the routes package.
+    """
+    st = jellyfin_receiver.status()
+    if not bool(st.get("enabled")):
+        raise HTTPException(status_code=503, detail="jellyfin integration disabled")
+
+    action = normalize_action(req.action, req.payload)
+    jellyfin_receiver.mark_command(action)
+    jellyfin_receiver.mark_heartbeat()
+    command_id = extract_command_id(req)
+    if is_duplicate_command(command_id):
+        return {"ok": True, "action": action or "unknown", "suppressed_duplicate_command": True}
+
+    try:
+        if action == "play":
+            source_url = (req.url or "").strip()
+            playlist_items = extract_playlist_items(req.payload)
+            item_ids = [it.get("id", "") for it in playlist_items if isinstance(it, dict)]
+            play_mode = extract_play_mode(req.payload)
+            if not source_url:
+                source_url = extract_play_url(req.payload)
+            item_id = extract_item_id(req.payload)
+            if not item_id and item_ids:
+                item_id = item_ids[0]
+            requested_item_id = str(item_id or "").strip()
+            media_source_id = extract_media_source_id(req.payload)
+            explicit_audio_idx = extract_audio_stream_index(req.payload)
+            explicit_sub_idx = extract_subtitle_stream_index(req.payload)
+            try:
+                settings_snapshot = state.get_settings()
+            except Exception:
+                settings_snapshot = {}
+            auth_token = access_token()
+            pref_audio_idx = ""
+            pref_sub_idx = ""
+            resolved_detail: dict[str, object] = {}
+            if item_id:
+                resolved_item = resolve_playable_item(item_id, media_source_id=media_source_id)
+                item_id = str(resolved_item.get("item_id") or item_id).strip()
+                resolved_detail = resolved_item.get("detail") if isinstance(resolved_item.get("detail"), dict) else {}
+                media_source_id = _first_nonempty_str([
+                    resolved_item.get("media_source_id") if isinstance(resolved_item, dict) else "",
+                    media_source_id,
+                ])
+                if requested_item_id and item_id and item_id != requested_item_id:
+                    if item_ids and item_ids[0] == requested_item_id:
+                        item_ids[0] = item_id
+                    if playlist_items and isinstance(playlist_items[0], dict) and str(playlist_items[0].get("id") or "").strip() == requested_item_id:
+                        playlist_items[0] = {
+                            **playlist_items[0],
+                            "id": item_id,
+                            "media_source_id": media_source_id or str(playlist_items[0].get("media_source_id") or "").strip(),
+                        }
+                pref_audio_idx, pref_sub_idx = preferred_stream_indices(item_id)
+                if not media_source_id:
+                    detail = resolved_detail if isinstance(resolved_detail, dict) else {}
+                    if not detail:
+                        try:
+                            detail = jellyfin_receiver.get_item_detail(item_id)
+                        except Exception:
+                            detail = {}
+                    media_source_id = _first_nonempty_str(
+                        [
+                            detail.get("media_source_id") if isinstance(detail, dict) else "",
+                            detail.get("MediaSourceId") if isinstance(detail, dict) else "",
+                        ]
+                    )
+            audio_stream_index = explicit_audio_idx or pref_audio_idx
+            subtitle_stream_index = explicit_sub_idx or pref_sub_idx
+            if not source_url and item_id:
+                source_url = build_item_stream_url(
+                    item_id,
+                    server_url=str(st.get("server_url") or ""),
+                    api_key=auth_token,
+                    media_source_id=media_source_id,
+                    audio_stream_index=audio_stream_index,
+                    subtitle_stream_index=subtitle_stream_index,
+                )
+            source_url = normalize_source_url(
+                source_url,
+                server_url=str(st.get("server_url") or ""),
+                api_key=auth_token,
+            )
+            source_url = apply_stream_params(
+                source_url,
+                audio_stream_index=audio_stream_index,
+                subtitle_stream_index=subtitle_stream_index,
+            )
+            source_url = apply_media_source_param(source_url, media_source_id=media_source_id)
+            if not media_source_id:
+                media_source_id = extract_media_source_id_from_url(source_url)
+            selected_stream: dict[str, str] = {"mode": "direct", "reason": "", "media_source_id": media_source_id}
+            if item_id:
+                selected_stream = select_playback_url(
+                    item_id=item_id,
+                    source_url=source_url,
+                    server_url=str(st.get("server_url") or ""),
+                    api_key=auth_token,
+                    media_source_id=media_source_id,
+                    audio_stream_index=audio_stream_index,
+                    subtitle_stream_index=subtitle_stream_index,
+                    settings=settings_snapshot,
+                )
+                source_url = normalize_source_url(
+                    str(selected_stream.get("url") or source_url),
+                    server_url=str(st.get("server_url") or ""),
+                    api_key=auth_token,
+                )
+                source_url = apply_stream_params(
+                    source_url,
+                    audio_stream_index=audio_stream_index,
+                    subtitle_stream_index=subtitle_stream_index,
+                )
+                media_source_id = _first_nonempty_str(
+                    [
+                        str(selected_stream.get("media_source_id") or "").strip(),
+                        extract_media_source_id_from_url(source_url),
+                        media_source_id,
+                    ]
+                )
+            if not source_url:
+                raise HTTPException(status_code=400, detail="play command requires url")
+            start_sec = extract_start_seconds(req)
+            try:
+                suppress_recent_stop = bool(
+                    getattr(player, "recent_jellyfin_stop_matches", lambda **_: False)(
+                        item_id=item_id,
+                        source_url=source_url,
+                        media_source_id=media_source_id,
+                    )
+                )
+            except Exception:
+                suppress_recent_stop = False
+            if suppress_recent_stop and (start_sec is None or float(start_sec) <= 1.0):
+                return {
+                    "ok": True,
+                    "action": "play",
+                    "suppressed_recent_stop_replay": True,
+                    "now_playing": state.NOW_PLAYING,
+                }
+            if should_suppress_duplicate_play(source_url, item_id, start_sec):
+                return {"ok": True, "action": "play", "suppressed_duplicate": True, "now_playing": state.NOW_PLAYING}
+            # If a play command explicitly asks to queue and we are already playing,
+            # add items to queue without interrupting current playback.
+            if play_mode in ("playnext", "playlast") and player.is_playing():
+                try:
+                    settings_snapshot = state.get_settings()
+                except Exception:
+                    settings_snapshot = {}
+                queued = []
+                existing_item_media: dict[str, set[str]] = {}
+                existing_urls: set[str] = set()
+                def _remember(iid_raw: object, url_raw: object, mid_raw: object = "") -> None:
+                    iid = canonical_item_id(iid_raw)
+                    if not iid:
+                        iid = extract_item_id_from_url(str(url_raw or ""))
+                    mid = canonical_media_source_id(mid_raw)
+                    if not mid:
+                        mid = canonical_media_source_id(
+                            extract_media_source_id_from_url(str(url_raw or ""))
+                        )
+                    if iid:
+                        mids = existing_item_media.setdefault(iid, set())
+                        mids.add(mid)
+                    qurl_existing = canonical_url_key(url_raw)
+                    if qurl_existing:
+                        existing_urls.add(qurl_existing)
+
+                _remember(
+                    (state.NOW_PLAYING or {}).get("jellyfin_item_id"),
+                    (state.NOW_PLAYING or {}).get("url"),
+                    (state.NOW_PLAYING or {}).get("jellyfin_media_source_id"),
+                )
+                for q in list(state.QUEUE):
+                    if isinstance(q, dict):
+                        _remember(q.get("jellyfin_item_id"), q.get("url"), q.get("jellyfin_media_source_id"))
+                    else:
+                        _remember("", q, "")
+
+                def _seen(iid_raw: str, qurl_raw: str, mid_raw: str = "") -> bool:
+                    iid = canonical_item_id(iid_raw)
+                    if not iid:
+                        iid = extract_item_id_from_url(qurl_raw)
+                    mid = canonical_media_source_id(mid_raw)
+                    if not mid:
+                        mid = canonical_media_source_id(extract_media_source_id_from_url(qurl_raw))
+
+                    if iid:
+                        seen_mids = existing_item_media.get(iid)
+                        if seen_mids:
+                            # Allow different known media-source variants of the same item.
+                            # Unknown/blank media source is treated as duplicate-safe and blocks additional variants.
+                            if not mid:
+                                return True
+                            if mid in seen_mids or "" in seen_mids:
+                                return True
+
+                    qurl = canonical_url_key(qurl_raw)
+                    if qurl and qurl in existing_urls:
+                        return True
+                    return False
+
+                source_for_queue = source_url
+                if source_for_queue:
+                    selected_queue = select_playback_url(
+                        item_id=item_id,
+                        source_url=source_for_queue,
+                        server_url=str(st.get("server_url") or ""),
+                        api_key=auth_token,
+                        media_source_id=media_source_id,
+                        audio_stream_index=audio_stream_index,
+                        subtitle_stream_index=subtitle_stream_index,
+                        settings=settings_snapshot,
+                    )
+                    source_for_queue = normalize_source_url(
+                        str(selected_queue.get("url") or source_for_queue),
+                        server_url=str(st.get("server_url") or ""),
+                        api_key=auth_token,
+                    )
+                    q_item = smart_item_from_url(source_for_queue)
+                    q_title = str(q_item.get("title") or "") if isinstance(q_item, dict) else ""
+                    q_channel = str(q_item.get("channel") or "") if isinstance(q_item, dict) else ""
+                    source_media_source_id = _first_nonempty_str(
+                        [
+                            media_source_id,
+                            q_item.get("jellyfin_media_source_id") if isinstance(q_item, dict) else "",
+                            extract_media_source_id_from_url(source_for_queue),
+                        ]
+                    )
+                    if not _seen(item_id, source_for_queue, source_media_source_id):
+                        queued.append(
+                            {
+                                "url": source_for_queue,
+                                "title": q_title or (f"Jellyfin item {item_id}" if item_id else "Jellyfin item"),
+                                "provider": "jellyfin",
+                                **({"channel": q_channel} if q_channel else {}),
+                                **({"thumbnail": q_item.get("thumbnail")} if isinstance(q_item, dict) and q_item.get("thumbnail") else {}),
+                                **({"jellyfin_item_id": item_id} if item_id else {}),
+                                **({"jellyfin_media_source_id": source_media_source_id} if source_media_source_id else {}),
+                                "jellyfin_stream_mode": str(selected_queue.get("mode") or ""),
+                                "jellyfin_stream_reason": str(selected_queue.get("reason") or ""),
+                            }
+                        )
+                        _remember(item_id, source_for_queue, source_media_source_id)
+                # Prefer rich playlist items when available.
+                for entry in playlist_items:
+                    iid = str(entry.get("id") or "").strip()
+                    if not iid:
+                        continue
+                    q_audio_idx, q_sub_idx = preferred_stream_indices(iid)
+                    if explicit_audio_idx:
+                        q_audio_idx = explicit_audio_idx
+                    if explicit_sub_idx:
+                        q_sub_idx = explicit_sub_idx
+                    qurl = build_item_stream_url(
+                        iid,
+                        server_url=str(st.get("server_url") or ""),
+                        api_key=auth_token,
+                        media_source_id=str(entry.get("media_source_id") or "").strip(),
+                        audio_stream_index=q_audio_idx,
+                        subtitle_stream_index=q_sub_idx,
+                    )
+                    if not qurl:
+                        continue
+                    selected_q = select_playback_url(
+                        item_id=iid,
+                        source_url=qurl,
+                        server_url=str(st.get("server_url") or ""),
+                        api_key=auth_token,
+                        media_source_id=str(entry.get("media_source_id") or "").strip(),
+                        audio_stream_index=q_audio_idx,
+                        subtitle_stream_index=q_sub_idx,
+                        settings=settings_snapshot,
+                    )
+                    qurl = normalize_source_url(
+                        str(selected_q.get("url") or qurl),
+                        server_url=str(st.get("server_url") or ""),
+                        api_key=auth_token,
+                    )
+                    if qurl == source_for_queue:
+                        continue
+                    q_media_source_id = str(entry.get("media_source_id") or "").strip()
+                    if _seen(iid, qurl, q_media_source_id):
+                        continue
+                    qtitle = str(entry.get("title") or "").strip() or f"Jellyfin item {iid}"
+                    queued.append(
+                        {
+                            "url": qurl,
+                            "title": qtitle,
+                            "provider": "jellyfin",
+                            "jellyfin_item_id": iid,
+                            **({"jellyfin_media_source_id": q_media_source_id} if q_media_source_id else {}),
+                            "jellyfin_stream_mode": str(selected_q.get("mode") or ""),
+                            "jellyfin_stream_reason": str(selected_q.get("reason") or ""),
+                        }
+                    )
+                    _remember(iid, qurl, q_media_source_id)
+                if queued:
+                    with state.QUEUE_LOCK:
+                        if play_mode == "playnext":
+                            state.QUEUE[:0] = queued
+                        else:
+                            state.QUEUE.extend(queued)
+                        qlen = len(state.QUEUE)
+                        queue_snapshot = list(state.QUEUE)
+                    try:
+                        state.persist_queue()
+                    except Exception:
+                        pass
+                    try:
+                        player.prime_mpv_up_next_from_queue(force=True)
+                    except Exception:
+                        pass
+                    try:
+                        lead = queued[0] if isinstance(queued[0], dict) else {}
+                        lead_title = str((lead or {}).get("title") or "").strip()
+                        queued_count = len(queued)
+                        if queued_count > 1:
+                            qtext = f"Queued {queued_count} items"
+                        else:
+                            qtext = f"Queued next: {lead_title or 'item'}"
+                        ui["toast"](
+                            text=qtext,
+                            duration=ui["notification_display_sec"](),
+                            level="info",
+                            icon="share",
+                            image_url=(lead.get("thumbnail") if isinstance(lead, dict) else None),
+                        )
+                    except Exception:
+                        pass
+                    emit_progress_hint()
+                    ui["queue_event"]("jellyfin_queue", queue=queue_snapshot, queue_length=qlen, source="jellyfin")
+                    ui["jellyfin_event"](
+                        "queue_only",
+                        refresh_active_tab=True,
+                        refresh_status=True,
+                        reason=play_mode,
+                    )
+                    return {"ok": True, "action": "queue_only", "queue_mode": play_mode, "queued": len(queued), "queue_length": qlen}
+            stopped_payload = None
+            if play_mode == "playnow":
+                cur = state.NOW_PLAYING if isinstance(state.NOW_PLAYING, dict) else None
+                if isinstance(cur, dict) and bool(player.is_playing()):
+                    cur_copy = dict(cur)
+                    cur_item_id = canonical_item_id(cur_copy.get("jellyfin_item_id"))
+                    cur_url_key = canonical_url_key(cur_copy.get("url"))
+                    next_item_id = canonical_item_id(item_id)
+                    next_url_key = canonical_url_key(source_url)
+                    replacing = False
+                    if cur_item_id and next_item_id:
+                        replacing = cur_item_id != next_item_id
+                    elif cur_url_key and next_url_key:
+                        replacing = cur_url_key != next_url_key
+                    if replacing:
+                        pos = None
+                        dur = None
+                        try:
+                            with player.MPV_LOCK:
+                                pos = player.mpv_get("time-pos")
+                                dur = player.mpv_get("duration")
+                        except Exception:
+                            pos = None
+                            dur = None
+                        stopped_payload = stopped_snapshot_from_now(cur_copy, pos, dur)
+            clear_queue_for_play = play_mode == "playnow"
+            playback_service.suppress_auto_next(2.0)
+            play_item_payload = smart_item_from_url(source_url, start_pos=start_sec)
+            play_target = play_item_payload if isinstance(play_item_payload, dict) else source_url
+            now = playback_service.play_now(
+                play_target,
+                use_resolver=bool(req.use_ytdlp),
+                cec=False,
+                clear_queue=clear_queue_for_play,
+                mode="jellyfin_play",
+                start_pos=start_sec,
+            )
+            # Preserve Jellyfin identifiers for progress/session reporting.
+            if isinstance(now, dict):
+                now = dict(now)
+                if isinstance(play_item_payload, dict):
+                    now = merge_playback_metadata(now, play_item_payload)
+                if item_id:
+                    now["jellyfin_item_id"] = item_id
+                    if media_source_id:
+                        now["jellyfin_media_source_id"] = media_source_id
+                    now["jellyfin_stream_mode"] = str(selected_stream.get("mode") or "direct")
+                    now["jellyfin_stream_reason"] = str(selected_stream.get("reason") or "")
+                    play_session_id = _first_nonempty_str([
+                        (req.payload or {}).get("play_session_id") if isinstance(req.payload, dict) else "",
+                        (req.payload or {}).get("PlaySessionId") if isinstance(req.payload, dict) else "",
+                    ])
+                    if play_session_id:
+                        now["jellyfin_play_session_id"] = play_session_id
+                    try:
+                        detail = jellyfin_receiver.get_item_detail(item_id)
+                    except Exception:
+                        detail = {}
+                    now = enrich_now_stream_metadata(
+                        now,
+                        detail=detail if isinstance(detail, dict) else {},
+                        audio_stream_index=audio_stream_index,
+                        subtitle_stream_index=subtitle_stream_index,
+                    )
+                playback_service.update_now_playing(now)
+            # Playlist-style play command support: enqueue remaining ItemIds.
+            if item_ids and len(item_ids) > 1:
+                extra_items = playlist_items[1:] if len(playlist_items) > 1 else [{"id": iid, "title": "", "media_source_id": ""} for iid in item_ids[1:]]
+                queued = []
+                seen_item_media: dict[str, set[str]] = {}
+                seen_urls: set[str] = set()
+
+                def _remember_seen(iid_raw: object, url_raw: object, mid_raw: object = "") -> None:
+                    iid = canonical_item_id(iid_raw)
+                    if not iid:
+                        iid = extract_item_id_from_url(str(url_raw or ""))
+                    mid = canonical_media_source_id(mid_raw)
+                    if not mid:
+                        mid = canonical_media_source_id(extract_media_source_id_from_url(str(url_raw or "")))
+                    if iid:
+                        seen_item_media.setdefault(iid, set()).add(mid)
+                    key = canonical_url_key(url_raw)
+                    if key:
+                        seen_urls.add(key)
+
+                def _seen(iid_raw: str, url_raw: str, mid_raw: str = "") -> bool:
+                    iid = canonical_item_id(iid_raw)
+                    if not iid:
+                        iid = extract_item_id_from_url(url_raw)
+                    mid = canonical_media_source_id(mid_raw)
+                    if not mid:
+                        mid = canonical_media_source_id(extract_media_source_id_from_url(url_raw))
+                    if iid:
+                        mids = seen_item_media.get(iid)
+                        if mids:
+                            if not mid:
+                                return True
+                            if mid in mids or "" in mids:
+                                return True
+                    key = canonical_url_key(url_raw)
+                    return bool(key and key in seen_urls)
+
+                _remember_seen(now.get("jellyfin_item_id") if isinstance(now, dict) else "", now.get("url") if isinstance(now, dict) else "", now.get("jellyfin_media_source_id") if isinstance(now, dict) else "")
+                for existing in list(state.QUEUE):
+                    if isinstance(existing, dict):
+                        _remember_seen(existing.get("jellyfin_item_id"), existing.get("url"), existing.get("jellyfin_media_source_id"))
+                    else:
+                        _remember_seen("", existing, "")
+
+                for entry in extra_items:
+                    iid = str(entry.get("id") or "").strip()
+                    if not iid:
+                        continue
+                    q_audio_idx, q_sub_idx = preferred_stream_indices(iid)
+                    if explicit_audio_idx:
+                        q_audio_idx = explicit_audio_idx
+                    if explicit_sub_idx:
+                        q_sub_idx = explicit_sub_idx
+                    qurl = build_item_stream_url(
+                        iid,
+                        server_url=str(st.get("server_url") or ""),
+                        api_key=auth_token,
+                        media_source_id=str(entry.get("media_source_id") or "").strip(),
+                        audio_stream_index=q_audio_idx,
+                        subtitle_stream_index=q_sub_idx,
+                    )
+                    if not qurl:
+                        continue
+                    q_media_source_id = str(entry.get("media_source_id") or "").strip()
+                    selected_q = select_playback_url(
+                        item_id=iid,
+                        source_url=qurl,
+                        server_url=str(st.get("server_url") or ""),
+                        api_key=auth_token,
+                        media_source_id=q_media_source_id,
+                        audio_stream_index=q_audio_idx,
+                        subtitle_stream_index=q_sub_idx,
+                        settings=settings_snapshot if isinstance(settings_snapshot, dict) else {},
+                    )
+                    qurl = normalize_source_url(
+                        str(selected_q.get("url") or qurl),
+                        server_url=str(st.get("server_url") or ""),
+                        api_key=auth_token,
+                    )
+                    q_media_source_id = _first_nonempty_str(
+                        [
+                            str(selected_q.get("media_source_id") or "").strip(),
+                            extract_media_source_id_from_url(qurl),
+                            q_media_source_id,
+                        ]
+                    )
+                    if _seen(iid, qurl, q_media_source_id):
+                        continue
+                    qtitle = str(entry.get("title") or "").strip() or f"Jellyfin item {iid}"
+                    queued.append(
+                        {
+                            "url": qurl,
+                            "title": qtitle,
+                            "provider": "jellyfin",
+                            "jellyfin_item_id": iid,
+                            **({"jellyfin_media_source_id": q_media_source_id} if q_media_source_id else {}),
+                            "jellyfin_stream_mode": str(selected_q.get("mode") or ""),
+                            "jellyfin_stream_reason": str(selected_q.get("reason") or ""),
+                        }
+                    )
+                    _remember_seen(iid, qurl, q_media_source_id)
+                if queued:
+                    with state.QUEUE_LOCK:
+                        state.QUEUE.extend(queued)
+                        queue_snapshot = list(state.QUEUE)
+                    try:
+                        state.persist_queue()
+                    except Exception:
+                        pass
+                    try:
+                        player.prime_mpv_up_next_from_queue(force=True)
+                    except Exception:
+                        pass
+                    ui["queue_event"]("jellyfin_playlist", queue=queue_snapshot, queue_length=len(queue_snapshot), source="jellyfin")
+            if isinstance(stopped_payload, dict) and stopped_payload:
+                emit_stopped_payload(stopped_payload)
+            emit_progress_hint()
+            ui["jellyfin_event"]("play", refresh_active_tab=True, refresh_status=True, reason=play_mode or "play")
+            return {"ok": True, "action": "play", "now_playing": now}
+
+        if action == "stop":
+            res = controls["stop"]()
+            return {"ok": True, "action": "stop", "result": res}
+
+        if action == "pause":
+            out = {"ok": True, "action": "pause", "result": controls["pause"]()}
+            emit_progress_hint()
+            return out
+
+        if action in ("resume", "unpause"):
+            out = {"ok": True, "action": "resume", "result": controls["resume"]()}
+            emit_progress_hint()
+            return out
+
+        if action == "seek":
+            sec = extract_seek_seconds(req)
+            if sec is None:
+                raise HTTPException(status_code=400, detail="seek command requires start_pos or payload.position")
+            out = {"ok": True, "action": "seek", "result": controls["seek"](float(sec))}
+            emit_progress_hint()
+            return out
+
+        if action == "next":
+            out = {"ok": True, "action": "next", "result": controls["next"]()}
+            emit_progress_hint()
+            return out
+
+        if action == "previous":
+            out = {"ok": True, "action": "previous", "result": controls["previous"]()}
+            emit_progress_hint()
+            return out
+
+        if action == "set_volume":
+            vol = extract_volume(req)
+            if vol is None:
+                raise HTTPException(status_code=400, detail="set_volume requires payload.VolumeLevel or payload.volume")
+            out = {"ok": True, "action": "set_volume", "result": controls["set_volume"](vol)}
+            emit_progress_hint()
+            return out
+
+        if action == "mute":
+            out = {"ok": True, "action": "mute", "result": controls["mute"](True)}
+            emit_progress_hint()
+            return out
+
+        if action == "unmute":
+            out = {"ok": True, "action": "unmute", "result": controls["mute"](False)}
+            emit_progress_hint()
+            return out
+
+        raise HTTPException(status_code=400, detail=f"unsupported jellyfin action: {req.action}")
+    except HTTPException:
+        raise
+    except Exception as e:
+        jellyfin_receiver.mark_error(str(e))
+        raise HTTPException(status_code=500, detail=f"jellyfin command failed: {e}")
 
 
 jellyfin_receiver.register_progress_provider(progress_snapshot)
